% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_gender.R
\name{age_trigger_map}
\alias{age_trigger_map}
\title{Age-based fraud trigger}
\usage{
age_trigger_map(
  claims_file,
  age_column_name,
  triggers_file,
  procedure_code,
  lower_age_limit,
  upper_age_limit
)
}
\arguments{
\item{claims_file}{data set containing claim-level information}

\item{age_column_name}{column within the claims file that indicates the age of the policy holder}

\item{triggers_file}{file containing the age ranges for each procedure code}

\item{procedure_code}{column in the triggers_file that contains the code/identifier for the medical procedure}

\item{lower_age_limit}{column in the triggers_file that contains the lower age limit for getting the procedure done}

\item{upper_age_limit}{column in the triggers_file that contains the upper age limit for getting the procedure done}
}
\value{
input claims_file with age_flag column
}
\description{
The triggers file contains reasonable age ranges for getting a certain medical procedure done. This function creates a flag to identify all the claimants whose age is falls outside this range.
             The age should be specified consistently in the claims_file and the triggers_file for this function to work.
}
\examples{

library(dplyr)
data(claims_data_sample)
data(trigger_data)

age_trigger_map(
claims_file = claims_data_sample,
age_column_name = "patient_age",
triggers_file = trigger_data,
lower_age_limit = "lower_age_limit",
upper_age_limit = "upper_age_limit",
procedure_code = "primary_procedure_code") \%>\%
group_by(patient_age) \%>\%
summarise(age_flag = mean(age_flag))

}
