% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim_count_amount.R
\name{claim_amount_trigger_map}
\alias{claim_amount_trigger_map}
\title{claim amount-based fraud trigger}
\usage{
claim_amount_trigger_map(
  claims_file,
  claim_paid_field,
  triggers_file,
  procedure_code,
  procedure_amount_field
)
}
\arguments{
\item{claims_file}{data set containing claim-level information}

\item{claim_paid_field}{column in the claims_file that contains the claim amount paid by the insurer}

\item{triggers_file}{file containing the agreed specific amount for each procedure}

\item{procedure_code}{column in the triggers_file that contains the code/identifier for the medical procedure}

\item{procedure_amount_field}{column in the triggers_file that contains the agreed specific amount for each procedure}
}
\value{
input claims_file with claim_amount_ratio and claim_amount_flag columns
}
\description{
For each procedure there is a specific amount that is generally agreed between the insurer and the medical service provider. This is especially the case for cashless claim processing. If the claim amount is higher than this agreed procedure specific amount it will raise a flag.
}
\examples{

library(dplyr)

data(claims_data_sample)
data(trigger_data)

claim_amount_trigger_map(
claims_file = claims_data_sample,
claim_paid_field = "approved_allowed_amount",
triggers_file = trigger_data,
procedure_code = "primary_procedure_code",
procedure_amount_field = "package_amount_trigger") \%>\%
select(c("approved_allowed_amount", "claim_amount_ratio", "claim_amount_flag"))

}
