% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claim_count_amount.R
\name{claim_count_trigger_map}
\alias{claim_count_trigger_map}
\title{Claim count-based fraud trigger}
\usage{
claim_count_trigger_map(
  claim_count,
  claim_count_pa,
  claim_count_trigger_file,
  procedure_code_field,
  claim_count_trigger
)
}
\arguments{
\item{claim_count}{file that contains the total number of times a policy holder claimed for a given procedure in a given time period}

\item{claim_count_pa}{column that contains the total number of claims per year by a policy holder}

\item{claim_count_trigger_file}{file containing maximum number of  times a policy holder can claim for each procedure}

\item{procedure_code_field}{column in the claim_count_trigger_file that contains the code/identifier for each medical procedure}

\item{claim_count_trigger}{column in the claim_count_trigger_file that indicates the number of times a procedure can be claimed by a policyholder before it can be considered as potential fraud}
}
\value{
input claims_file with claim_count_flag column
}
\description{
We don't expect the policyholder to get a given medical treatment more than few times a year, depending on the nature of the treatment. If the policy holder has taken a given treatment unreasonably high number of times, it needs to be investigated for fraud. This the function identifies all such policyholders.
}
\examples{

library(dplyr)

data(claims_data_sample)
data(trigger_data)

claim_count_trigger_map(
claim_count = claims_data_sample,
claim_count_pa = "claim_count_pa",
claim_count_trigger_file = trigger_data,
procedure_code_field = "primary_procedure_code",
claim_count_trigger = "claim_count_trigger") \%>\%
group_by(claim_count_pa)\%>\%
summarise(claim_count_flag = mean(claim_count_flag))

}
