% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claims_data_sample}
\alias{claims_data_sample}
\title{Sample Claims data file}
\format{
`claims_data_sample`
}
\usage{
claims_data_sample
}
\description{
Sample claims dataset containing different fields
}
\keyword{datasets}
