% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{close_prox_trigger_map}
\alias{close_prox_trigger_map}
\title{Close proximity-based fraud trigger}
\usage{
close_prox_trigger_map(
  claims_file,
  treatment_start_date,
  policy_commencement_date,
  triggers_file,
  procedure_code,
  close_prox_days
)
}
\arguments{
\item{claims_file}{data set containing claim-level information}

\item{treatment_start_date}{column in the claims_file that contains the starting date of the treatment}

\item{policy_commencement_date}{column in the claims_file that contains the commencement date of the policy}

\item{triggers_file}{file containing the number of days for each procedure within which we can call it as a close proximity claim}

\item{procedure_code}{column in the triggers_file that contains the code/identifier for the medical procedure}

\item{close_prox_days}{column in the triggers_file that contains the number of days for each procedure within which can call it as a close proximity claim}
}
\value{
input claims_file with close_prox_flag and claim_duration_days columns
}
\description{
a claim is in close proximity if the treatment start date is very close to the policy commencement date. This function  will raise a flag for such claims.
}
\examples{

library(dplyr)

data(claims_data_sample)
data(trigger_data)

close_prox_trigger_map(
claims_file = claims_data_sample,
treatment_start_date = "treatment_start_date",
policy_commencement_date = "policy_commencement_date",
triggers_file = trigger_data,
procedure_code = "primary_procedure_code",
close_prox_days = "close_prox_days") \%>\%
select(c("treatment_start_date", "policy_commencement_date",
"claim_duration_days", "close_prox_flag"))

}
