% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hosp_days_trigger_map}
\alias{hosp_days_trigger_map}
\title{Hospital admitted days-based fraud trigger}
\usage{
hosp_days_trigger_map(
  claims_file,
  no_of_days_stayed,
  triggers_file,
  procedure_code,
  admission_days_trigger
)
}
\arguments{
\item{claims_file}{data set containing claim-level information}

\item{no_of_days_stayed}{column in the claims_file that contains the number of days actually admitted in the hospital}

\item{triggers_file}{file containing the reasonable number of days a claimant could be admitted in the hospital for each procedure}

\item{procedure_code}{column in the triggers_file that contains the code/identifier for the medical procedure}

\item{admission_days_trigger}{column in the triggers_file that contains the reasonable number of days a claimant could be admitted in the hospital for each procedure}
}
\value{
input claims_file with no_days_ratio and hosp_days_flag columns
}
\description{
For each procedure there is a reasonable number of days a claimant could be admitted in the hospital.If the number of days actually admitted is higher than what is reasonable for that procedure this function will raise flag.
}
\examples{

library(dplyr)

data(claims_data_sample)
data(trigger_data)

hosp_days_trigger_map(
claims_file = claims_data_sample,
no_of_days_stayed = "no_of_days_stayed",
triggers_file = trigger_data,
procedure_code = "primary_procedure_code",
admission_days_trigger = "admission_days_trigger") \%>\%
select(c("no_of_days_stayed", "admission_days_trigger", "hosp_days_flag"))

}
