% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hospital_distance_trigger_map}
\alias{hospital_distance_trigger_map}
\title{Hospital distance-based fraud trigger}
\usage{
hospital_distance_trigger_map(
  claims_file,
  residence_location_field,
  hospital_location_field,
  hospital_distance_file,
  residence_location_map,
  hospital_location_map,
  hospital_distance_field,
  distance_threshold
)
}
\arguments{
\item{claims_file}{data set containing claim-level information}

\item{residence_location_field}{column in the claims_file with the location of the policy holder}

\item{hospital_location_field}{column in the claims_file with the location of the hospital}

\item{hospital_distance_file}{file with reference table of distances between places in the data}

\item{residence_location_map}{column in the hospital_distance_file  with the list of all possible residence locations}

\item{hospital_location_map}{column in the hospital_distance_file with all the possible hospital locations}

\item{hospital_distance_field}{column in the hospital_distance_file that contains the distance between the hospital location and the residence location}

\item{distance_threshold}{distance level beyond which claim is flagged for fraud}
}
\value{
input claims_file with hospital_distance_flag column
}
\description{
It is reasonable to expect that the policyholder gets treated in the nearest hospital. If the distance between the residence location and the hospital location is more than a defined threshold, this function will raise a flag
}
\examples{

library(dplyr)

data(claims_data_sample)
data(location_distance)

hospital_distance_trigger_map(
claims_file = claims_data_sample,
residence_location_field = "residence_location",
hospital_location_field = "hospital_location",
hospital_distance_file = location_distance,
residence_location_map = "location_1",
hospital_location_map = "location_2",
hospital_distance_field = "distance",
distance_threshold = 100) \%>\%
select(c("residence_location", "hospital_location", "hospital_distance_flag"))

}
