% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hospital_empanelled_trigger_map}
\alias{hospital_empanelled_trigger_map}
\title{Empanelled hospitals(medical service providers)-based fraud trigger}
\usage{
hospital_empanelled_trigger_map(
  claims_file,
  hospital_id_field,
  empanelled_hospitals_list,
  empanelled_hospital_id
)
}
\arguments{
\item{claims_file}{data set containing claim-level information}

\item{hospital_id_field}{column with unique identity number of the hospitals}

\item{empanelled_hospitals_list}{file containing the list of all the empanelled hospitals}

\item{empanelled_hospital_id}{column in the empanelled_hospital_list file that contains all the unique hospital ids}
}
\value{
input claims_file with hospital_empanelled_flag column
}
\description{
The insurer generally empanells hospitals to service its policy holders. In the process of empanelment the insurer ensures that the hospital has the required facilities and also agrees the tarrif for each of the treatments. If the hospital mentioned is not a part of the empanelled list of hospitals it will raise a flag
}
\examples{

library(dplyr)

data(claims_data_sample)
data(hospital_list)

hospital_empanelled_trigger_map(
claims_file = claims_data_sample,
hospital_id_field = "medical_service_provider_id",
empanelled_hospitals_list = hospital_list,
empanelled_hospital_id = "hosp_id") \%>\%
group_by(medical_service_provider_id)\%>\%
summarise(hospital_empanelled_flag = mean(hospital_empanelled_flag))

}
