% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_distance}
\alias{location_distance}
\title{Location Distance Data}
\format{
`location_distance`
}
\usage{
location_distance
}
\description{
Distances between two locations used in the triggers
}
\keyword{datasets}
